+++ packages/expo-ui/CHANGELOG.md
@@ -31,6 +31,7 @@
 - [iOS] Add `font` modifier ([#40553](https://github.com/expo/expo/pull/40553) by [@nishan](https://github.com/intergalacticspacehighway))
 - [iOS] Add `getItemAsync` to `ShareLink` ([#40391](https://github.com/expo/expo/pull/40391) by [@intergalacticspacehighway](https://github.com/intergalacticspacehighway))
 - [iOS] Add Grid component and modifiers `gridCellUnsizedAxes`, `gridCellColumns`, `gridColumnAlignment`, `gridCellAnchor` ([#40485](https://github.com/expo/expo/pull/40485) by [@doombladeoff](https://github.com/doombladeoff))
+- [iOS] Add `labelView` to LabeledContent component ([#](https://github.com/expo/expo/pull/) by [@focux](https://github.com/focux))
 
 ### 🐛 Bug fixes